pyproject.toml
[tool.poetry]
name = "fastapi-sn"
version = "0.1.0"
description = "Social network project using FastAPI and GoLang"
authors = ["NelttjeN <tatyapa0880@yandex.ru>"]
readme = "README.md"
packages = [{include = "fastapi_sn"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.101.1"
pydantic = "^2.2.1"
psycopg2-binary = "^2.9.7"
sqlalchemy = "^2.0.20"
alembic = "^1.11.3"
asyncpg = "^0.28.0"
python-dotenv = "^1.0.0"
python-json-logger = "^2.0.7"
pyjwt = "^2.8.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = "3.3.0"
python-multipart = "^0.0.6"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
flake8-bugbear = "^23.7.10"
flake8-builtins = "^2.1.0"
flake8-commas = "^2.1.0"
flake8-functions-names = "^0.4.0"
flake8-import-order = "^0.18.2"
flake8-length = "^0.3.1"
flake8-quotes = "^3.3.2"
flake8-variables-names = "^0.0.6"
flake8_simplify = "^0.20.0"
pep8-naming = "^0.13.3"
coverage = "^7.3.0"
pyflakes = "^3.1.0"
pipdeptree = "^2.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
